+++ front-end/src/index.js
@@ -8,16 +8,13 @@ import App from "./App";
 import "antd/dist/antd.min.css";
 import "./index.css";
 
-
-
 const store = legacy_createStore(reducers, compose(applyMiddleware(thunk)));
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <Provider store={store}>
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>
+    <React.StrictMode>
+      <App />
+    </React.StrictMode>
   </Provider>
 );
-