+++ front-end/Changelog.md
@@ -9,5 +9,6 @@
 -Create add,update and view ui for the book 
 -Create add,update and view ui for the author
 -integrated redux and create add author function
+- Create author backend integration
 
 ____________Book______________________